Vim: shorten column macro
As noted in the last commit, I suspect I could make this macro simpler
by changing the regex to deal better with leading whitespace. This is
the result. It's still not perfect (for example, it makes no attempt to
save and restore the original tab-stop setting), but it works.

Note: the trailing whitespace on the first line is significant.

  vnoremap <leader>zc
    \:s/\v^ +//g<CR>       " remove leading whitespace
    \gv:s/\v  +/\r/g<CR>   " reselect last selection, split columns onto lines
    \gvj:sort<CR>          " reselect last selection, sort
    \V}k:!column -c 78<CR> " select region, columnify
    \:set tabstop=8<CR>    " set tabstops to match `column` executable
    \V}:retab<CR>          " select region again; convert tabs to spaces
    \gv>:noh<CR>           " indent, and remove search highlighting

Signed-off-by: Wincent Colaiuta <win@wincent.com>