Vim: fix "Invalid expression" errors in <leader>r mapping
With a recentish Vim update my <leader>r mapping would occasionally bomb
out with:

  ={ '00': 'r', '01': 'no', '10': ''}[&rnu . &nu]
  E716: Key not present in Dictionary: 11
  E15: Invalid expression: { '00': 'r', '01': 'no', '10': ''}[&rnu . &nu]

Somehow, we were getting into a state where both 'nu' and 'rnu' were
active at the same time.  The cause here is that Vim gained the ability
to show the current line number on the "0th" line (ie. the current line)
when relative numbering is active; the two settings can now be toggled
independently of one another.

In reality, the "rnu + nonu" combination isn't very interesting at all,
so we want to cycle between:

  rnu + nu
  nornu + nu
  nornu + nonu

Note this means that to cycle it is not longer sufficient to perform a
single `set` command; we may actually have to perform two. So, in the
interests of simplicity, always perform two, which in turn means that
our little dictionary lookup trick becomes a little less terse. Oh well.

With this, we're totally robust against all possible combos of "&nu .
&rnu" (ie. 00, 01, 10, 11), so we should never see those pesky E716 and
E15 errors again.

Additionally, given that 'rnu' and 'nu' can be set independently we
explicitly set both, rather than one or the other, in the settings.vim
file.

Signed-off-by: Wincent Colaiuta <win@wincent.com>