Git: use a ridiculously obvious interactive prompt color
It's possible that this is even too strong (I might want to swap out the
"reverse" for "ul" if this proves to be too much of an eyesore; but note
that "bold" is next to useless on this terminal), but I find it tiresome
when working my way through a long run of hunks in `git add --patch` to
have to visually scan upwards to where the previous hunk ends and the
current one begins.

This makes the prompts so strident that it is impossible to not see
them, and makes it very clear where the hunks begin and end.

Tested with Solarized light and dark.

Signed-off-by: Wincent Colaiuta <win@wincent.com>