Git: lower help.autocorrect interval
When I originally set this to 10 (ie. 1 second) I wanted an interval
that was long enough to notice, but not too long to become annoying, and
at the same time just long enough that I could Control-C if I saw that
it was about to do something really stupid that I wanted to abort.

Having had the setting in place for a couple of years now, I doubt I've
even tried to abort a command more than half-a-dozen times (and that is
a generous estimate).

So, I'm lowering the interval. 0.2 seconds is definitely long enough to
notice that something happened, but not so long that you have time to
think, "Oh come on, hurry up and move on now".

Signed-off-by: Wincent Colaiuta <win@wincent.com>