Add install script for setting up terminfo entries
I've had to set this up on three machines in the last few days, and this
basic process has worked on all of them; I'm going to call this good
enough to "reify" in the form of a script.

One gotcha: if you're doing a static tmux build you'll probably wind up
with a $HOME/share/terminfo directory. You should do that before you run
this script.

Unlikely to work everywhere, but it is working for me on OS X 10.9 and a
couple of 5.x CentOS boxen.

Signed-off-by: Wincent Colaiuta <win@wincent.com>