Vim: allow for per-host overrides file
I thought I was going to need this in order to get UltiSnips working on
two different machines, one with a Vim that links against Python 2 and
another linking against Python 3.

It turns out that with the latest UltiSnips, everything works fine on
both anyway. Nevertheless, I think this is a useful feature to have
around, so keeping it in.

Signed-off-by: Wincent Colaiuta <win@wincent.com>