Vim: unset CommandTMaxHeight
I'm working on an order-of-magnitude larger project now and limiting the
match listing to only 10 rows is making it harder to find stuff.

Removing this setting makes it revert to its default value of 0, which
means "use as much height as is available and necessary" to display all
the results.

Signed-off-by: Wincent Colaiuta <win@wincent.com>