Vim: teach :Ack/:Lack to be more tolerant
For a long time now I've been using :Ack gingerly, knowing that certain
things will cause it to freak out once the args are passed to the shell.
This includes even innocuous things like:

  :Ack save!

This change transforms that internally to:

  ack [opts] 'save!'

I can even do things like:

  :Ack -i \bFoo_?Bar

The downside? I can't search for things with spaces in them. I am not
sure if this is a big deal -- although I suspect it is -- so I am going
to sit on this for a while before pushing it.

Signed-off-by: Wincent Colaiuta <win@wincent.com>