*sh: handle tmux/SSH edge case
I was noticing that every now and again my lovely ~/.ssh/ssh_auth_sock
symlink was getting replaced with a useless circular reference back to
itself.

This would happen any time I was inside a tmux session and ran a tmux
command directly from the command-line. (This isn't something I do
everyday, but it's not entirely rare either; I'm not talking here about
doing anything tricky like nesting sessions, but rather innocuous stuff
like running `tmux ls`).

This commit adds a check to make sure we only create the symlink if the
thing we're linking to isn't already a symlink (as is the case when we
create the circular references).

Signed-off-by: Wincent Colaiuta <win@wincent.com>