Slate: rework aliases for tenkeyless keyboards
For a long time I used external keyboards with numeric keypads, which
meant that many of these bindings wouldn't work when I was on my laptop
alone.

Now that I've switched external keyboards at work (to a tenkeyless
board), they don't work even there.

This commit reworks the bindings to be accessible even on the laptop.
Fewer options mean getting more down with a smaller number of keys, at
the expense of having to lean more heavily on the `chain` operator,
which can mean having to cycle a bit more.

Note: I am not sure this will work as expected because I can't test the
effect of multiple `next` declarations in a single `chain` expression. I
am hoping it is subbed in at evaluation time and not at run time, so
that a rule like:

  chain A | B | C | A next | B next | C next

is equivalent to the following, assuming screen numbers 1 and 2:

  chain A 1 | B 1 | C 1 | A 2 | B 2 | C 2

and not:

  chain A 1 | B 1 | C 1 | A 2 | B 1 | C 2

Will just have to test it out on when I get back to work.

Signed-off-by: Wincent Colaiuta <win@wincent.com>