Slate: fix ordering issues
In commit 4b2b9a2bfe00e I added some new mappings so that I could access
all of my Slate configuration with a tenkeyless keyboard. It turns out
that the fear I noted in that commit message was borne out:

  chain A | B | C | A next | B next | C next

is indeed equivalent to the following, given screen numbers 1 and 2:

  chain A 1 | B 1 | C 1 | A 2 | B 1 | C 2

One work around is to write the config like this:

  chain A next | B | C

and that works, but it is not a very nice user experience. For example,
say I want to push a window over the left on the main display, I have to
push the mapping 4 times to get it where I want.

Instead, the approach this commit takes is to explicitly identify the
monitors, preferring the main display if present:

  chain A 1 | B 1 | C 1 | A 2 | B 2 | C 2

This means that I can easily push a window over to the left on the main
display with a single key combination. Pushing something on the
alternate display is a little less pleasant (4 key combos), but as this
is less common it might be ok.

Short of switching to a JS config (which I was planning to do anyway),
this is about the best we can do for now.

Signed-off-by: Wincent Colaiuta <win@wincent.com>