Vim/Git: replace git-jump with vcs-jump
Port git-jump to Ruby and teach it to handle both Mercurial and Git. The
code is a bit of a hack job (blend of Ruby-style and the original
Shell-plus-Perl-style) but it seems to handle the basic cases ok.

For now, "diff" is supported for Mercurial and Git, but "merge" and
"grep" are only supported for Git.

I would have just adapted the existing script, but Mercurial's `hg diff`
has no equivalent of `git diff --relative` that I am aware of, and
manipulating ("relativizing") the paths in Bash would have been a little
more painful. Switching to Ruby, where we have the Pathname class, makes
path manipulation easier.

As a nicety, if the editor is Vim, we open the quickfix window instead
of only jumping to the first error.

On errors, such as when we don't seem to be inside a repo, we raise an
exception; this means that you see some diagnostic info on the console
when running `vcs-jump` on the command line. When running it inside Vim
using the mapping, we redirect error output to /dev/null, which means
you just get an empty error listing (ie. nothing happens).

Signed-off-by: Wincent Colaiuta <win@wincent.com>