Slate: port stock config to JS format
I used to use a 104-key keyboard (ie. with numeric keypad) and so was
able to keep some of my less frequently used Slate bindings on the
keypad. With the move to a tenkeyless board, I've had to rework my
bindings to make heavy use of chains (see commits 4b2b9a2bfe00e and
1fabf0ea7181b), but the results were suboptimal.

For example, to push a window to the left two-thirds of my laptop
display, even if the window was already on the laptop, would require me
to press a binding as many as 6 times to cycle through to the end of the
chain; ie. if the window was on the right of the laptop display:

1. Move to left half of main display
2. Move to left third of main display
3. Move to left two-thirds of main display
4. Move to left half of laptop
5. Move to left third of laptop
6. Move to left two-thirds of main display.. victory!

(It's true that the typical case was nowhere near as bad, but still, it
is a noteworthy degenerate case.)

With the move to JS, we can implement literally any logic that we
desire.

This commit is a straight port, but adds the following enhancements:

  - the cycle always starts on the current screen
  - after 1 second of inactivity, any given cycle resets
  - switching from one chain to any other resets as well

This is still not 100% perfect, but it's a big improvement over what
we had before. (The one edge case that isn't dealt with here yet is if
you switch apps, or windows within an app, before the one second timeout
has passed; the chain will continue without resetting; I should be able
to at least partly address this in a follow-up commit.)

Signed-off-by: Wincent Colaiuta <win@wincent.com>