Slate: teach chain to reset on changing windows
The comments pretty much say it all. This gets us 99% of the way there,
I think, and is probably good enough for now. Hopefully the list of
"difficult" apps in the `fingerprint()` function won't grow too long; if
it does, I think I'll just drop the title across the board.

Signed-off-by: Wincent Colaiuta <win@wincent.com>