Slate: kill off a comment
I'm making this as a separate commit so that in cleaning out the cruft
from the tree, I have a place (ie. this commit) to explain my plan of
action.

As noted in that comment, the "Google Chrome Canary" rule doesn't match
any more. (The app on disk is still called "Google Chrome Canary.app",
but at runtime it presents itself as "Google Chrome".)

I'd still very much like to be able to target it though. It's possible
that by passing in a custom function for the "Google Chrome" target, we
can distinguish between the two seeing as we'll have a window reference
and from there can get a PID, and we also have the option of shelling
out to an arbitrary command.

Signed-off-by: Wincent Colaiuta <win@wincent.com>