Vim: add <leader>p mapping to show current path
I'm currently connected to an Apple Cinema Display with a full-screen
terminal window and a pretty small font. Even so, with my Vim display
cut into 11 splits (because of the number of files I'm juggling right
now), I sometimes want to know the path of the current file and the
status line is truncated so I can't see it.

Enter, <leader>p, and its companion, <leader>pp, which copies the
current path into the clipboard (via Clipper).

Signed-off-by: Wincent Colaiuta <win@wincent.com>