Vim: try to detect iTerm via ITERM_SESSION_ID as well
If you connect to an SSH server that has ITERM_SESSION_ID whitelisted in
the AllowEnv settings in sshd_config, then your remote Vim instance can
detect that it is running inside iTerm.

Checking for this as well as ITERM_PROFILE increases our chances of
getting detected without having to resort to the dot-file signal hack.

Signed-off-by: Wincent Colaiuta <win@wincent.com>