Vim: fix bug in Ack command related to spaces
We were missing some escaping here, which meant that queries like:

  :Ack foo\ bar baz

(ie. find "foo bar" in the directory "baz) were equivalent to:

  ack "foo bar baz"

because all spaces, both "\ " and " ", were being treated indistinctly.
This made it not only impossible to scope the search within a particular
directory, but it also made it very unlikely that any matches would be
found (unless the "foo bar" your looking for happens to be followed by
" baz").

With this change, they are now once again correctly distinguished, which
means that searches scoped to specific directories can work once again,
as can searches with options; eg:

  :Ack -w boop

Signed-off-by: Wincent Colaiuta <win@wincent.com>