Vim: add easydir plug-in
It's probably overkill to create a submodule and a plug-in with
documentation etc for about 7 lines of VimScript (which is so simple
it's never likely to be updated), but anyway...

I wasn't sure if this was a good idea or not (what if typos lead you to
create unwanted hierarchies?) but given that I mostly open stuff with
Command-T, tab-completion and other tricks (like `%` expansion, and my
`<leader>e` binding), I doubt I'll run into problems too often. But I do
expect this to save me some time and the annoyance of having to do an
explicit `:!mkdir %<tab>` etc.

Signed-off-by: Wincent Colaiuta <win@wincent.com>