Vim: improve autocompletion with YouCompleteMe and supertab
I am not sure if I'm going to like this (YouCompleteMe ends up
auto-displaying the completion menu a _lot_), or if it is going to be
too slow (I already have too many plugins...), but I'm going to give it
a try.

The experience of expanding UltiSnips snippets is pretty slick with
this: `<tab>` to expand snippets, and also to jump to the next
placeholder. I was rather amazed to see that even `<s-tab>` works in the
terminal (for jump backwards).

For compliance with my muscle memory, I've got `<c-j>` and `<c-k>` bound
to cycling through items in the completion menu.

For reference, if I want to turn off the as-you-type pop-up, I'll be
able to do that with:

  let g:ycm_auto_trigger = 0

Signed-off-by: Wincent Colaiuta <win@wincent.com>