Vim: make UltiSnipsEdit work with new snippets dir
In b01906f I renamed "UltiSnips" to "ultisnips". If you try to use
":UltiSnipsEdit" with that configuration, UltiSnips will try to edit a
file at `.vim/UltiSnips` (which works on OS X only because HFS
filesystems are usually formatted as case-insensitive; it doesn't work
elsewhere).

If you instead use ":UltiSnipsEdit!", you get a list of files to chose
from:

  1: /home/glh/code/wincent/.vim/UltiSnips/all.snippets
  2: /home/glh/code/wincent/.vim/UltiSnips/javascript.snippets
  3: /home/glh/code/wincent/.vim/ultisnips/javascript.snippets

This isn't awesome. Luckily, by switching to `g:UltiSnipsSnippetsDir`
instead, we get the desired behavior. I was originally confused by this
when reading the docs, but it appears that this is the one I'm supposed
to be using. (`UltiSnipsSnippetDirectories` is for adding additional
folders in which to search for snippets, so you can pull down other
repos of snippets, for instance; `UltiSnipsSnippetsDir` is for defining
the one directory where you want your own snippets to live.)

Signed-off-by: Wincent Colaiuta <win@wincent.com>