Vim: fix UltiSnips on case-sensitive filesystems
So I thought I had it all figured out in 44186362d0976c3ab9d4887ef3ae66,
but it turns out that things were only working on OS X, with its
case-insensitive file system; on Linux, I could _edit_ snippets but I
couldn't use them.

So, it's not one or the other setting which is needed, but rather both
of them. `g:UltiSnipsSnippetsDir` tells UltiSnips where to find files
for editing (`:UltiSnipsEdit`), and `g:UltiSnipsSnippetDirectories`
tells it where to find them for snippet expansion.

Signed-off-by: Greg Hurrell <greg@hurrell.net>