Slate: only position first Chrome window
Chrome (and I think OS X) usually does a pretty good job of positioning
new windows:

- if there is no window in the top-left corner, place new window there
- if there is already a window there, place the new window on top, but
  slightly displaced down and to the right
- repeat for new windows

Other constraints:

- if the window is full height, displace only to the right
- if the window is full height and width, don't displace it
- each new window defaults to the size of the previous window

This means that the code I had in here was a little bit maddening. The
first Chrome window would open on the right half of the screen, and the
second window would open on the left half (due to Chrome or OS X's
logic) and then jump over to the right (due to Slate).

This commit makes it so that only the first visible window is moved by
Slate. This works pretty well for the most part because I usually have
only one Chrome window visible at a time. It does mean that the `repeat`
setting in our layout config is not as useful as before, but such is
life.

Signed-off-by: Greg Hurrell <greg@hurrell.net>