Vim: tweak search highlighting
I find the garish yellow to be very hard to look at; it gives me an
almost uncontrollable urge to hit `:noh`, even if I am in the middle of
a search and know I'll have to hit `n` again within milliseconds.

Hopefully this blue will be easier to stomach. Seems to look good in
both light and dark Solarized variants.