Vim: make backspace (mostly) move tab-wise
The docs for this setting are a bit confusing (see `:h 'sts'`), but the
effect is as follows:

Prior to this commit, shifting or using tab/backspace in the leading
whitespace of a line would move us forward/backward by '`shiftwidth`',
but doing so in the trailing whitespace would only do so for tabbing,
not backspacing.

With this commit, backspacing in trailing space moves by `'shiftwidth'`
as well, provided you've used tab immediately prior to backspacing. In
other words:

- if you press `<space>` 10 times at the end of a line, `<bs>` moves
  spacewise
- if you press `<tab>` 10 times at the end of a line, `<bs>` moves
  tabwise
- if you press `<tab>` 5 times then `<space>` 5 times, `<bs>` moves
  spacewise (including once you move past the 5 spaces)
- if you press `<space>` 5 times and then `<tab>` 5 times, `<bs>` moves
  tabwise (including once you move past the 5 tabs)
- if you enter non-whitespace characters at the end of a line, `<bs>`
  moves spacewise

Partly stolen from: https://github.com/lencioni/dotfiles/commit/f48d3508