Vim: add vim-projectionist and some config
I haven't really felt the need for this until now, but working on small
Node projects makes me want to be able to jump back and forth between
test files.

@tpope does it again, with vim-projectionist. Adding it as a submodule
with some sample configuration that allows you to jump back and forth
between:

- src/thing.js
- src/__tests__/thing-test.js

Will likely add more configuration subsequently, but this small piece
removes the single largest pain point I was feeling.

Note that I originally targeted this to projects with a "package.json"
file at the root, but I'm going to start with something more general
("*") and see if it bites me.

* .vim/bundle/vim-projectionist 0000000...c5fda43 (110):
  > Ignore non-dictionary projections