Fix missing submodule entry for vim-projectionist
A search through my history reveals I did this in totally the wrong way:

  $ history -10000|grep projectionist
  6442  git clone https://github.com/tpope/vim-projectionist.git .vim/bundle/vim-projectionist
  6444  git add .vim/after/plugin/projectionist.vim .vim/bundle/vim-projectionist/

And funnily, Git kind of does the right thing... in the superproject, it
shows it as a submodule (probably because this is exactly the way
submodules used to be implemented; ie. with a full-blown `.git`
directory inside the submodule).

Fixed thusly:

  $ rm -rf .vim/bundle/vim-projectionist
  $ git rm .vim/bundle/vim-projectionist
  $ git submodule add https://github.com/tpope/vim-projectionist.git .vim/bundle/vim-projectionist

At which point we have the desired entry in `.gitmodules`, and `.git` in
the submodule is now a "link" file containing:

  gitdir: ../../../.git/modules/.vim/bundle/vim-projectionist