Vim: fix uncomment mapping side effect
Not sure why the old mapping did this, but it was trying to fold in
addition to doing what I wanted it to do (remove a comment).

Setting it up in an alternative way (which is probably the way I should
have done it originally anyway) makes things work without the undesired
side effect.