Vim: suppress bell
The bell is particularly annoying when doing things like hitting "j" at
the bottom of a buffer, or any number of other not-very-error-ish
"errors".

The trick is to turn _on_ the visual bell, but unset the terminal code
for actually producing the bell.

Accoding to the tip here:

http://vim.wikia.com/wiki/Disable_beeping

this shouldn't work for gvim, because the value of "t_vb" gets reset to
its default value _after_ reading the vimrc and on starting the GUI.
However, that doesn't seem to apply for MacVim, where this change
appears to work just fine, so I am not going to worry about applying any
workaround at this point; if I ever need to, it should be trivial to
apply the same setting in the gvimrc file.