Vim: Tweak status line
Lots of small tweaks here:

- make sure there is a blank space between the right and left halves of
  the status line
- make right half size flexible; it doesn't need to be 35 characters
  wide
- collapse various markers (eg. "[RO]", "[+]", "[javascript]" etc) into
  one (eg. "[+,RO,javascript]")
- highlight those same markers using italics
- show file encoding if it's set and not UTF-8
- use "%p" to show percentage, which is simply numerical, rather than
  the "%P" which I previously used
- use shorter line output (eg. "10/100" instead of "10 of 100")
- moved truncation point so that buffer number is always shown, even
  with very long file names
- split up statusline setting over multiple lines and put trailing
  comments on same lines (note there is one set of markers that are
  inside an item group, so have to be on the same line)

Overall, these changes give me a bit more room for filenames, and
by making use of some functions I was able to avoid the undesired switch
to capital letters that would have come with the move from "%y" to "%Y".

Vim is remarkably helpful here, collapsing the item group away to
nothing if all of the functions and markers return nothing, and even
stripping the leading "," that happens if one of the custom functions
returns a non-empty string and is also the first thing in the item group
to return something.

Note that I would have preferred to use script-local functions for the
custom functionality, but the usual `s:foo` approach doesn't work here.
Setting up a couple of "autoload" functions (conveniently in the same
file) does give me some degree of namespacing, at least... (See
http://stackoverflow.com/a/17184285 for more on that.)

The autocmd that sets up the User1 highlight is very Solarized-specific,
but oh well... Also note that I had to swap the order of the background
setting in my `<leader>w` mapping to ensure the background is already
set when the autocmd runs.