Homebrew: switch from `brew bundle` to Brewdler
`brew bundle` was deprecated and then removed, with similar
functionality now residing in "brewdler"[0]. It's kind of ridiculous for
this thing to describe itself as "Bundler for non-Ruby dependencies from
Homebrew" seeing as it doesn't support locking or versioning, which is
the defining characteristic of Bundler, but anyway...

I considered just prefixing every line in my `Brewfile` with `brew ` and
forgetting about Brewdler, but the `brew brewdle dump` functionality is
useful, and will make keeping my `Brewfile` up-to-date easier (including
custom build arguments). There's also a `cleanup` feature that might be
handy, although I am not sure I see myself using it.

Created a separate `update-homebrew.sh` script to make bootstrapping and
updating the bundle straightforward (you still need to ensure Homebrew
is installed on your system before running it).

[0]: https://github.com/Homebrew/homebrew-brewdler