Vim: don't freak out when terminal window is narrow
When the terminal window is narrow, the output of `:hi` may contain
newlines, which when fed back into `:hi` cause it to barf (because the
subsequent lines are treated as distinct commands, but they are not
valid, generally being of forms like "guibg=...").

Make sure we always return one-line strings from the `ItalicizeGroup`
function.