Manage OS X defaults via Ansible
I thought about just shelling out the script, but the nice part about
YAML-izing this is that we get output from Ansible about what has
changed, whenever there is a change.

It is a bit verbose though (the items get dumped), so it would be nice
to suppress that. I fear it's not going to be possible though without
hacking Ansible.