Vim: ignore dirty worktree in Command-T submodule
Recent changes in Command-T have it dirtying one of its files when you
install it ("ruby/command-t/metadata.rb:). Let's update the
".gitmodules" configuration to keep this change out of the `git status`
output. Note that the submodule still shows up if the `HEAD` has
diverged from the expected value.