Vim: add VISUAL support to ** snippet
`**<tab>` behaves as before, leaving you in insert mode with the cursor
inside the newly created comment.

With a visual selection, `<tab>**<tab>` deletes the selection, inserts
the comment template, pastes in the selection you had previously, and
leaves you in insert mode at the end of the former selection.

Doesn't work awesomely for multi-line visual selections, but this is a
start at least.