Vim: Use prettier 'colorcolumn' format
Instead of one column, paint _all_ the columns to the right of the
limit. This looks pretty. If you're on a large display with a small
font, you will see only 255 columns get painted irrespective of the
limit you supply; this is an internal Vim limit.

As before, apply this only to the active window, to make it obvious
which one is active.

Pro-Tip via @clentfort. He does warn that this slows down redraws, so I
am going to take this for a test drive and see if it's worth it.