Vim: remove vim-orgmode
I played with it for a while, and it is pretty nice, but I got multiple
Python tracebacks during my brief experimentation with it, so I decided
it probably wasn't as robust as I'd like it to be.

Also, integration with my iCloud-based reminders across devices is
pretty important for me, so I decided to give 2Do a try (it's not Vim,
but it works).

Reverts 1885d3.

Keeping vim-speeddating though, which I added as a vim-orgmode
dependency.