Vim: switch to `g:CommandTTraverseSCM = 'dir'`
This always gets me (to the point that I wonder if the default of "file"
might be wrong).

- "file" looks for the nearest SCM root from the current file and uses
  that as the base
- "dir" looks for the nearest SCM root from the current working
  directory and uses that as the base

Command-T itself is very project centric, and encourages people to
anchor Vim in the top level of the project directory, so I think "dir"
is the right choice.

Example scenario causing confusion:

- start off in your project directory
- `vim`
- `:h YCM`
- `:CommandT`

I personally would expect to see my project files at this point, but
instead I see YCM's, because the help document is in its directory (in
`~/.vim/bundle/YouCompleteMe` which happens to be a Git submodule, so
gets picked up at the appropriate base).

No more!

If I want to find something relative like that with Command-T, I'll
just have to `:cd ~/.vim/bundle/YouCompleteMe`.