Vim: fix trouble opening files with spaces using Command-T
I hadn't noticed this because:

- I strenuously avoid having filenames with spaces
- On the rare occasions I'm working in project that is nested inside a
  path containing a space, there was no problem because Command-T uses
  relative paths wherever possible
- Command-T itself takes pains to escape spaces; this is an issue
  deriving from my custom `GotoOrOpen` command and function

It turns out that `<f-args>` does crazy things with the passed in
values (see `:h <f-args>`). This might be helpful in the general case,
but it wreaks havoc with pre-escaped input, like I was using.
Specifically:

> A backslash followed by a character other than whitespace or a
> backslash remains unmodified.

So, given a carefully escaped input like "foo\\ bar.txt", it would get
mangled to "foo\", "bar.txt"; or something like that... I am not even
sure. The end result is that we'd issue a command like `:e foo bar.txt`
and it would complain with `E172: Only one file name allowed`.

The fix here is to switch to `<q-args>` instead. This means we get a
single argument and have to split it (maybe I didn't have to do that,
but it's the first thing I tried).

The `\zs` in the `split()` pattern tells Vim to start the pattern at
that point, which means "e " (the command) is not considered part of the
match, and so it doesn't get thrown away.