Vim: add a mapping for deleting lines from the quickfix listing
I like to use the quickfix listing as a working set of files I need to
modify. Sometimes I just want to get ignorable files (ie. irrelevant, or
already handled) out of the way.

This commit adds a `dd` mapping to delete a line from the quickfix. If
you try this manually (ie. with `:set modifiable` and `dd`), you'll find
the links in the listing get broken (they point at the wrong files).
This solution doesn't suffer from that problem, although it does jump us
back to the top of the list.

Note that the `d` visual mode mapping that I've included doesn't work
properly. Puzzlingly, my function is getting called once for each line
in the selection instead of once, which makes the `while` loop invalid.
Inconveniently, we're called in increasing order too, so we can't just
delete a line at a time.

Later I might consider adding support for motions in normal mode (eg.
`dj` to delete current line and line after it).

I'd also like to address the jumping to the top; we should probably only
jump up one entry, or potentially down one.