Vim: Fix YouCompleteMe double-tap woes
In order to get the double-tap thing sorted out, I've had to fork
YouCompleteMe and add a new setting to it
(`g:ycm_key_list_accept_completion`).

I'm pushing this commit first, so that I can show an example of use in
the PR that I send upstream, before I update the submodule.

As noted in the comment inline, I'm using the otherwise-unused `<F18>`
here in order to avoid a disambiguation delay with emmet.vim's mappings,
which all start with `<C-y>`.