Vim: add command mode mappings for auto-inserting \v
Sigh... I know the argument is that providing a way to make "very magic"
regexen the default would supposedly break a lot of plug-ins (ie. the
ones that don't specify an explicit `\{m,M,v,V}` in their patterns), I
just wish we had a non hacky way to do this, but we don't.

This gives us three new mappings which auto-expand to include `\v`:

  %s/     -> %s/\v
  g/      -> g/\v
  v/      -> v/\v

Note the included check to make sure these only kick in at the start of
the line; you wouldn't want a command like: `:e testing/foo` to expand
to `:e testing/\vfoo`, for example.

The one downside is the fact that these are multi-key mappings means
there's a slight delay before expansion occurs. In my brief play with
them, it doesn't seem too annoying though.

Not included: mappings to make visual selections followed by `:s` work
in the same way, because I don't really want to assume that `:s` is
always going to be `:substitute` and not the beginning of something
else.