Vim: Remove <leader>da mapping
I haven't needed to use this in a long time. This was added back in the
days when I was using the minibufexpl plug-in, and found that
long-running Vim sessions led to an unsightly accumulation of buffer
names stealing valuable real estate at the top of the screen, so I'd
periodically clean them out.

Nowadays, I simply don't care. The more buffers, the better (for better
autocomplete), and navigating them is not a problem with Command-T, no
matter how many of them there are.