Vim: add better "very magic slash" handling
This is the first step towards handling "/"-as-very-magic_slash for
range cases. In this commit, we have parity with the separate mappings
that I previously added before, but it's strictly better because we're
using `<expr>`, which means we don't see a brief flash on the
command-line showing the expression body, and it's not a multi-char
mapping so it's instant.

Next step will be teaching the new `mappings#very_magic_slash` function
to handle some special range cases. (Not all of them, because I don't
feel like implementing a regex parser...)