Vim: remove cursorcolumn shenanigans
Not only is `'cursorcolumn'` slow, it's a bit ugly too. I don't think I
need it anymore because of late I've made the display simpler and more
focused, and I've also added an auto-`zz` on jumping between search
results, which makes it a lot easier to see where the current match is.