Vim: Make Loupe cope with multiple windows better
`matchadd()` adds to the current window only, so on switching to another
window and trying to hit `n`, were were dying, because we saw the
`s:hlmatch` variable, but our `matchdelete()` call would blow up.

Make the variable window-specific to avoid this problem.

Still to handle: `<leader>n` in any window, should clear up the
window-specific variables for all windows.