Vim: clean up straggling matches
Fixes a minor polish issue. Repro:

- in one window, search for "foo"
- in another, hit `<leader>n` (to effectively do `:nohiglight`)
- go back to first window, the old match (added by `matchadd()`) is
  still in effect, even though search highlighting is off (ie. you can
  see `IncSearch` highlight whenever the cursor is within the match)

I originally wanted to fix this by looping over all windows and killing
off all matches on `<leader>n`; something like:

    function! loupe#private#clear_highlight() abort
      for l:tabpage in range(1, tabpagenr('$'))
        for l:window in range(1, tabpagewinnr(l:tabpage, '$'))
          let l:loupe_hlmatch = gettabwinvar(a:tabpage, a:window, 'loupe_hlmatch')
          if !empty(l:loupe_hlmatch)
            call matchdelete(l:loupe_hlmatch)
            call settabwinvar(a:tabpage, a:window, 'loupe_hlmatch', 0)
          endif
        endfor
      endfor
    endfunction

Alas, this doesn't work because `matchdelete()` only works in the
current window. We could try something hacky with `:windo`, but that
would likely cause the screen to flicker, and we'd have to worry about
iterating over each page separately, because `:windo` only operates on
the current tab page.

The approach taken here, then is to do this via an autocommand, and
inspect the value of `v:hlsearch`. If we have a lingering window-local
variable around, but search highlighting is off, we clean up.