tmux: don't wrap copy-mode searches
I find searching in the scrollback buffer to be super confusing because
of the way the position is displayed in reverse order; eg:

[0/12012] <-- means you're scrolled to the bottom
[12012/12012] <-- scrolled to the top

This is the opposite of what my mental model (which is based on line
numbers in a file) would expect, so when you end up bouncing between a
couple of results:

[5670/12012] <-- somewhere in the middle
[5570/12012] <-- somewhere in the middle

I really have no idea where I am. Let's anchor ourselves by at least
stopping when we can go no further up or down.