Vim: take the best parts of vim-vinegar
I tried vim-vinegar and wanted to like it, but the vast bulk of its
functionality is unused or doesn't work when you try to use it in
conjunction with NERDTree (specifically, when NERDTree is set to hijack
netrw).

So, make `-` work to show NERDTree and move up the directory hierarchy.