tmux: Switch prefix to space
Took me a while to figure out why this wasn't working... Mouse Locator
was swallowing `C-Space`. Turned off the keyboard shortcut for now, and
will be able to turn off the entire thing with the next version of OS X,
which has similar functionality built in anyway (jiggle the mouse to
magnify).

Doing this for consistency with my new Vim leader, and because typing
some of the tmux combos is more difficult since switching to Colemak
(and for the same reasons).

Considered keeping `a` around as a secondary prefix, in order to
accommodate occasional lapses of muscle memory, but then figured I may
as well go cold turkey so that I can benefit from naked `C-a` to jump to
the start of the command line, increment numbers in Vim etc...