Vim: be slightly smarter about opening folds on `:loadview` etc
On looking at this code I saw it could be smarter. Instead of opening
all folds, we can target only the range that matters (ie. the one with
the cursor in it).

For the new codepath, that means the current line, because the
`:loadview` command has already run at that point.

For the old codepath, it is the value of the `'"` mark.