Vim: start tracking spellfile
Previously I had this under a custom path, and I kept it out of version
control for reasons that I can't now remember.

Let's get it into version control, seed it with a few entries, and just
use the standard Vim path(s) by unsetting `'spellfile'` and letting the
default kick in.