tmux: Make `next-layout` binding work again
When I switched my prefix to Space, I replaced the `next-layout` binding
with one that sends the prefix (ie. `C-Space`).

Turns out, that's not useful to me at all[*], unlike the old prefix
(`C-a`), which was critical to me. So, undo that, and restore the
`next-layout` to its rightful place.

[*]: As proof that I literally never used it, consider that that key
combo used to be monopolized by Mouse Locator, until I deactivated it a
day or two ago.