vim: don't check for capital letters at start of sentences
This kind of error is obvious enough that I don't need the machine to
point it out for me, especially because it has the habit of pointing it
out in a lot of places where I don't want it to be considered an error
(eg. I don't ever want "tmux" to be capitalized, even if it is at the
start of a sentence).