tmux: refine scrollwheel behavior
Via Pro-Tip(tm) from @cHoco at:

https://github.com/tmux/tmux/issues/168#issuecomment-150963592

This fixes my issue where "inertial" scroll wheel events where taking me
in and out of copy mode instead of reliably exiting it at the bottom of
the copy mode scrollback buffer.

Additionally, because it has alternate screen detection, and just sends
the {Up,Down} keys through in that case, this makes the scrollwheel work
in `less` (and therefore `man`) as well, which is pretty awesome. (Note
that I dropped the `-X` flag to `less` in order to make this work; it
needs to send terminal control sequences to enter/exit alternate screen
mode, so it means that I'm back to having `less` clear on exit... I
think it's probably worth it.)