*sh: Prevent `less` from hiding small files
I recently dropped `X` from `$LESS` (enabling the alternate screen, so
that I can use the mousewheel in `less` when in tmux), but kept `F`
(which would exit immediately for small files that fit entirely on the
screen).

This combo meant that you wouldn't see short files at all.

Here's the quick fix (removing `F` as well, everywhere). This means
nothing will be invisible, but a bunch of other stuff is annoying (like
`git status` that hangs around until quit).

Better/deeper fix forthcoming, but wanted to get this in anyway to show
the intermediate state.