Hammerspoon: add window-level event monitoring
As noted in the comments, you currently have to do this a very verbose
way, but will eventually be able to rely on the `hs.window.filter`
module instead. Only not doing so now because of the warning in the
docs:

> WARNING: EXPERIMENTAL MODULE. DO NOT USE IN PRODUCTION.
> This module is for testing purposes only. It can undergo breaking API
> changes at any point and without notice.

Using a debug-level logger for now during development; will turn it back
down to "info" once I'm done.