hammerspoon: beware errors raised for window created events
This happens 100% of the time trying to access `element:id()` (or any
other method) on this video when switching to full screen mode:

https://www.facebook.com/StarWars/videos/1059652867419647/

The error appears to be an unexpected message to `nil` somewhere inside
the `id()` method. `element` itself is _not_ `nil`. Funnily,
`log.i(element)` successfully prints something, at least:

> hs.window: Greg Hurrell - Greg Hurrell shared Star Wars's video.
> (0x7fd253f99078)