hammerspoon: double check before responding to screen events
It seems the system can report screen change events even when nothing
substantive has changed, triggering an unwanted layout event. I've seen
this a few times although I haven't identified the exact cause.

I first suspected that showing/hiding the Dock or the menu bar might be
responsible, but that doesn't seem to be the case. I have noticed that
the unwanted events appear to happen mostly when moving the mouse in a
browser window, but that could just be because I happen to be in browser
windows a lot.

Anyway, I've been running with this fix for a couple of days and it
seems to do the trick, so, time to commit it. It should make sure that
events and layout changes only trigger it the number of displays, or
their resolutions, change.