vim: deal with com.apple.quarantine
I dug deeper and found out that Marked 2 is setting these things:

http://support.markedapp.com/discussions/questions/8670

The app will need an obscure entitlement
(`com.apple.security.files.user-selected.executable`) set:

http://stackoverflow.com/questions/9544874/how-can-i-stop-my-app-from-setting-the-quarantine-bit#comment36131063_11993894

Looks like a fix will be out in the next version, but in the meantime,
this commit applies a workaround.

Perhaps I should be using `system` instead of `!` here, but punting for
now.