Hammerspoon: avoid spurious layout changes
Even when only one monitor is plugged in, I am seeing "layout" changes
even when no resolution change has taken place. I have a couple of
theories for bad citizens that may be causing this (like Flash video
plug-ins), but in any case, I can partly mitigate this by ignoring
changes whenever the screen count is unchanged at 1.

I added logging to catch these changes red-handed, and this is the case
I am trying to prevent here:

```
20:36:31 ** Warning:   wincent: got screen event 1 -> 1
20:36:31 ** Warning:   wincent: layout changed! 2560x1440@1x -> 1440x900@2x
```

You can see here it is claiming to flip in and out of Retina mode here,
which is clearly not happening. I may loosen this check even further in
the future to trigger layout only on screen number changes, ignoring
dimensions. That is probably good enough.