vim: set up Haskell autocompletion
ff1d1fe intalled neco-ghc which is used to get Haskell autocompletion
features. e3e752d cherry-picked a bug fix from the upstream
YouCompleteMe repo that allows me to tune down my heavy-handed
`ycm_filetype_specific_completion_to_disable` workaround.

This commit sets things up as per the instructions on:

https://github.com/eagletmt/neco-ghc

(Not exactly "as per", as I've expanded the list of semantic triggers a
bit so that I don't have to hit `C-x C-o` as often to force the
omnicomplete to kick in.)

Requires ghc-mod, which I had a bit of trouble installing because it
wanted "happy" (a parser generator), but "happy" wouldn't install
without the `PATH` being set correctly, so the overall process was
something like:

```
$ cabal install ghc-mod # boom
$ cabal update
$ cabal install ghc-mod # boom
$ cabal install happy # boom
$ cabal update # boom, wtf? DNS is borked, again...
$ sudo launchctl unload -w /System/Library/LaunchDaemons/com.apple.mDNSResponder.plist
$ sudo launchctl load -w /System/Library/LaunchDaemons/com.apple.mDNSResponder.plist
$ cabal update # yay
$ export PATH=$PATH:~/.cabal/bin
$ cabal install happy # yay
$ cabal install ghc-mod # yay
```