vim: add `iabbrev` set-up
I've been aware of this approach (using `iabbrev` to catch common typos)
for a while, but didn't want to use it for fear of it reinforcing bad
habits (or at least, not giving me the impetus for reforming bad habits).

This was especially true because in 2015 I switched to Colemak from
Qwerty, cold-turkey, so I wanted to be hyper-aware of my mistakes so I
could correct them.

Nevertheless, when I stumbled across the "vim-fat-finger" plug-in, which
has a list of over 4.3k abbreviations(!)[0], I realized that a small,
personalized set of abbreviations like @lencioni has[1] is actual quite
moderate and nothing to be afraid of.

So, let's start. I think my initial policy here is going to be to add
abbreviations for words which I mistype when I'm moving fast, not by
hitting the wrong keys, but by hitting in the wrong order. So,
"funciton" makes it in. Funnily, the way @lencioni tends to mistype this
same word is "functino", even though we're both using Colemak. The
"vim-fat-finger" plugin corrects "fucntion" but neither of the other
mentioned variants.

[0]: https://github.com/chip/vim-fat-finger/blob/master/plugin/fat-finger.vim
[1]: https://github.com/lencioni/dotfiles/blob/master/.vim/plugin/abbreviations.vim