vim: use vim-abolish a bit more
I thought of it for all of 10 seconds after committing 678912d6d and
decided that I'll lean on vim-abolish and dispense with my manually
`:iabbrev` entirely. Now I'm going to eagerly await the arrival of new
typos that are amenable to automated correction (alas, not all are) so
that I can further populate this file.

The first new arrival is `varient` (`variant`), in addition to
`funciton` (`function`) that I am moving (and augmenting) in this
commit.