vim: put the cursor keys to good use
Inspired by:

  http://howivim.com/2016/damian-conway/

Who uses the cursor keys to move through the quick fix listing like
this:

- `<Left>` is `:cnext` (`<Left><Left>` is `:cnfile`)
- `<Right>` is `:cprevious` (`<Right><Right>` is `:cpfile`)

I am not a huge fan of repeated-key mappings, due to the delay, so I'm
taking a slight variant of this: the `<Down>` and `<Up>` keys are mapped
to `:cnext` and `:cprevious`, and the `<Right>` and `<Left>` keys are
mapped to `:cnfile` and `:cpfile`.

Thanks to the SpaceFN layout, this means I can navigate through quickfix
result listings very fast by holding space and hitting any of J, K, L,
I (Qwerty keys) or N, E, I, U (Colemak equivalents in the same
physical position). I use the quickfix listing literally dozens of times
a day, and obviously use `:cn`, `:cp`, `:cnf`, `:cpf` even more, so this
is going to save me a lot of time.