vim: Work around long-standing gripe with NERDTree alternate file
It has long annoyed me that once I've started to explore the directory
of a file using my `-` mapping, I can't easily get back to the file with
`<space><space>` (ie. `<leader><leader>`) or `<C-^>`. The same issue
affects vim-vinegar (although I am not using right now). There is a
workaround when using netrw, which is setting the `g:netrw_altfile`
option, which causes netrw to use `:keepalt` under the covers:

    https://github.com/tpope/vim-vinegar/issues/25

Anyway, I've added a similar option to NERDTree in my fork and have an
open PR for it here:

    https://github.com/scrooloose/nerdtree/pull/550

* roles/dotfiles/files/.vim/bundle/nerdtree 4ebbb53...9843fd3 (1):
  > Make window creation command configurable