vim: shuffle around some mappings using <LocalLeader>
Seeing as `<Space>` is my `<Leader>`, I am free to use `\` for something
else. Let's abuse `<LocalLeader>` for that purpose. Normally this would
be for file-type-specific bindings, but I am going to use it for "global
but not as frequently accessed". Sounds insane when you say it out loud
("local" is the new "global"), but I think it will work well.

- <Leader>c (syntax coloring fixup, rare) is moved off to
  <LocalLeader>c.
- <Leader>c, now freed up, gets reassigned to `:CommandTCommand`.
- <Leader>h goes from `:help` to `:CommandTHelp`.
- Added a bunch of other mappings for different Command-T finders
  (line, tag, history, search etc).