Add maxfiles role
Sick of running into file descriptor limits. The defaults on OS X are
insanely low, and Apple changed something in the last couple of months
that means I'm seeing limit problems all the time (this corroborated by
talking to other people who have also started seeing this in recent
months).

This causes everything to fail in mysterious ways. Chrome crashing?
Probably open file limit. DNS lookups failing? Probably mDNSResponder
dead due to open file limit. (Etc...)

Note that this is our first use of `become`, so we need to get a sudo
password at start. I could conditionalize this on _which_ role(s) are
getting run, but I don't run this often enough to make that optimization
worthwhile.

After reboot, can verify with:

    launchctl limit
    sysctl -a|grep -E 'proc|files'
    ulimit -a

Note that the "hard limit" on open files (10240, defined as `OPEN_MAX`
in /usr/include/sys/syslimits.h) can apparently be overridden.

The limit for maxproc, however, seems to be hardcoded at 1064. I can set
it higher and `launctl` will claim it is such, but any attempt to go
higher with `ulimit -u` and we get:

    ulimit:1: value exceeds hard limit

Perhaps that is encoded inside Zsh, or it got compiled in when I built
Zsh.

See also:

- http://unix.stackexchange.com/questions/108174/how-to-persist-ulimit-settings-in-osx-mavericks
- http://docs.basho.com/riak/latest/ops/tuning/open-files-limit/