vim: Tweak Loupe highlighting
Prior to this commit all matching search results were highlighted with
the "Search" highlight group, and the current one was highlighted with
"IncSearch". This is ever so slightly annoying because "Search" is a
little more obvious than "IncSearch". This becomes a real problem,
though, under the "grayscale" color scheme, which I've been playing with
briefly tonight, as the two highlight groups look almost identical.

So, add a function to copy highlights, which in turn enables us to:

- Copy "Search" to some other group, which we're calling
  "LoupeHighlight".
- Link "Search" to a less obvious "VisualNOS".
- Tell Loupe to use "LoupeHighlight" (ie. the old "Search" styling) for
  the current match.

This works well in both "grayscale" and my prefered "ocean" color
scheme.