keyboard: tweak F-key config on the basis of tests
I had a couple of typos in here, and needed to swap the brightness
targets otherwise I ended up controlling the wrong display.

The FN stuff doesn't work though. I can see the FN modifier toggling on
and off in the event viewer but it doesn't actually modify the behavior
of the F-keys.