vim: Don't let Q go into Ex mode
I avoided doing this for a long time because I didn't want to put
blinders on myself to the utility of Ex mode before I really understood
it. Now, after years of Vim use, I have never intentionally entered Ex
mode nor needed it. It truly is a relic of a forgotten age; sorry to be
so callous.