keyboard: make F13 a lock
Just discovered that it is next to impossible to type
Control-Command-Option-F1 (my binding for activating screen layouts) on
the Realforce, because I have to hold F13 down at the same time. This is
much easier on the laptop, where it is just
Fn-Control-Command-Option-F1, and all those modifiers are right next to
each other.

So, let's turn F13 into a lock instead. Karabiner provides a helpful
overlay in this case so I can see when it is on.