vim: fix MacVim-only issues with statusline syntax coloring
No idea why this happens, but in MacVim at boot time:

- `hlID` returns a non-zero value.
- `synIDtrans` returns a non-zero value.
- `synIDattr` returns an empty string.

This in turn leads to `E254` ("Cannot allocate color {name}") when we try to set
up the highlight group.

After boot time, though, it works fine (ie. if you switch away and back,
triggering `FocusGained`, everything works fine).

Fix here is to pass extra mode parameter to `synIDattr`, which causes it to
return a non-empty string all the time. Still works in the terminal, naturally.