vim: add new <Leader>e mapping to change all instances
Current `<Leader>e` mapping is rarely used anymore now that I have `-` bound to
show me the current directory, and one of my common use cases is to save a
similar file, but I'm just doing `:sav %<tab>` for that now instead.

So, demote the old `<Leader>e` to `<LocalLeader>e`, and repurpose `<Leader>e` to
change all instances of the word under the cursor. This is something I did
manually quite a bit with `:%s/<C-R><C-W>/...`, but this is much easier. I
prefer it to the `c*` mapping because it allows me to see the string that I am
changing rather than blowing it away before I start inputting the new string.

Idea from:

https://www.reddit.com/r/vim/comments/4gjbqn/what_tricks_do_you_use_instead_of_popular_plugins/d2iatu9