vim: use <tab> to toggle folds
I originally set this up in c58e19ed276c but then backed it out in b934f55ca16ae
because I didn't want to lose the functionality of <c-i> (jump list navigation).

Decided I was sick of this limitation, and that I should use Karabiner to hack
around it. So, keep <tab> as is, but make <c-i> actually send <F6> when in
Terminal/Vim, and set up a mapping to catch it and reroute it to the real <c-i>.
Works in Vim (inside and outside of tmux) and MacVim. I would have preferred to
use a not-real function key, like F19, say, but I couldn't get that to work. Oh
well, I doubt I will miss the pre-existing functionality of F6 (if there even is
any... I wouldn't know).

One thing to note: I'm targeting `KeyCode::L` here because that's the physical
key I need to intercept. This is where the 'I' key is in the Colemak keyboard
layout. If you where doing this on another keyboard layout you would need to
modify accordingly.