vim: enable custom Markdown autocompleter
Updates my YouCompleteMe fork[0].... to point[1] to my new ycmd fork[1]... which
includes a new Markdown completer[2] to make writing links in my wiki[3] easier.
Depends on a custom config file[4] and the configuration in this commit in order
to work, but what it does work.

Hacks upon hacks upon hacks, but this is really cool.

* roles/dotfiles/files/.vim/pack/bundle/start/YouCompleteMe a2c8c09...3429617 (1):
  > Point ycmd at my fork

[0]: https://github.com/wincent/YouCompleteMe
[1]: https://github.com/wincent/YouCompleteMe/commit/3429617d6ca341e59e091d4217
[1]: https://github.com/wincent/ycmd
[2]: https://github.com/wincent/ycmd/commit/174a14ef9750dd089ee05613fccad961cc1
[3]: https://github.com/wincent/masochist/tree/content
[4]: https://github.com/wincent/masochist/blob/bfabfe98fd48620f747b8/.ycm_extra_conf.py