vim: fix weirdness with Error highlighting under 'termguicolors'
Problem: in terminal (both inside and outside tmux), the color of "Red" was
varying depending on whether it was used as a bg or fg color. Probably for
unfathomable reasons alluded to in `:h cterm-colors`, where it says:

> The number under "NR-16" is used for 16-color terminals ('t_Co' greater than
> or equal to 16).  The number under "NR-8" is used for 8-color terminals
> ('t_Co' less than 16).  The '*' indicates that the bold attribute is set for
> ctermfg.  In many 8-color terminals (e.g., "linux"), this causes the bright
> colors to appear.  This doesn't work for background colors! Without the '*'
> the bold attribute is removed. If you want to set the bold attribute in a
> different way, put a "cterm=" argument AFTER the "ctermfg=" or "ctermbg="
> argument. Or use a number instead of a color name.

Whatever. This fixes it. Now things look right everywhere (tmux, term, MacVim).