tmux: Make active/inactive panes more obvious
Use the same colors as we do in Vim to indicate focus/blur.

Note that when you have Vim and the terminal visible at the same time,
this change makes the "active"/focused color much more obvious, so I am
also implementing an EndOfBuffer change (which will take effect in Vim
8) to make the active part of the active buffer stand out even more.
With this, the rest of the screen should be pretty subdued.