zsh: add global R alias (mnemonic: "[r]eset")
What I really want is a binding that I can use to scroll back to the
most recently issued command in tmux scroll-back buffer, but in the
absence of that, this is a stop-gap measure.

Use case: working through screenful's of Flow errors, I always want to
fix the topmost error then re-run to see how that affects the output.
Finding where it all starts in the scrollback is a pain. So, instead,
you loop like this:

  R flow # then scroll up, fix an issue
  R flow # then scroll up, fix an issue

The other alternative is this:

  flow --color=always|less # fix an issue
  flow --color=always|less # fix an issue

But the problem with that is that I can only see the total error count
if I scroll down, and that's actually the first thing I always want to
see (because I want to see that the change I made actually made the
error count go down from 27 to 26).

In the long run, I will try to hack together the desired binding,
because this workaround is somewhat of a "baby-out-with-the-bathwater"
type solution, seeing as it nukes the scrollback.