vim: Make <CR> mapping work better on first try
If you hadn't executed a macro (eg. `@q`) then repeating the "last"
macro (ie. with `@@`) would just error out with error E748.

Ideally, I'd have a way to find out which register you last recorded
into and just use that instead of `@@`. That might be tricky to pull off
though, so for now I'm just going to assume that if `@@` doesn't work,
you probably recorded into the "q" register, which is by far the most
common, and use that.