hammerspoon: Implement custom Enter behavior
I am going to be forced to upgrade to Sierra soon, and Karabiner
Elements isn't ready yet, so seeing how much of it I can recreate via
Hammerspoon event taps. Code is very ugly at this point -- no
abstraction, and probably buggy too -- but my goal for now is just to
show that it is possible at all.

In this commit I add support for:

- Enter: tap to Enter
- Enter: press and hold to repeat Enter
- Enter: chord with another key to be Control plus that key

Temporarily have some table-printing code in there that I got off of
Stack Overflow or somewhere because Lua is so ridiculously minimal that
it doesn't have any thing like that, and it's useful for debugging
during development.