hammerspoon: favor rollover for fast key-presses
Based on the timing information I've been logging, it is very common for
the user to rollover when using a modifier, so let's bias towards
assuming that for everything faster the .2s. That works pretty well for
the Caps Lock/Control/Delete side; not so sure about return though. I
tend not to do fast roll-overs on that side, so it might not be optimal
there. Maybe by introducing per-key threshholds I can have my cake and
eat it too.