Revert "zsh: eliminate last unwanted global from `colors` file"
Ok, so it turns out that zsh anonymous functions don't work like
closures. This only seemed to work because I had `BASE16_CONFIG` still
in the global scope. What actually happens in practice is that, as a
local, it exists the first time `color` is called, but the next time
around it is gone.

This reverts commit 46a67e6e5a746c29a16f0439b95a95a886ac5143.