vim: Turn off Command-T "isengard" engine to shave 100ms off startup
I'm not actively working on this right now, so I may as well avoid the
unnecessary cost of starting up the daemon process.

Note that when I resume development of it, I can make this cheaper so
that turning it on won't be a regression (at the moment we are eagerly
loading the daemon; we just need to lazily load it on first use, or use
a `CursorHold`/`CursorHoldI` trick to load it pre-emptively but
non-blockingly).