mutt: move to first blank line in composer
When replying to a post with a large quoted body, we would jump down
into it.

We can't specify a fixed height to get us below the headers either, as
the number can vary (eg. hitting "a" to reply all on a message, where
the height depends on how many "all" means).