mutt: fix possible infinite loop in work filters
Main issue is we were mutating a table during iteration. We were also
relying on `differential` being in scope and previously initialized
(which it usually is, but may not be up to date).

Also get rid of redundant check for "[Abandoned]" in the subject. It
does look a bit slow though, so I might have to put it back.