mutt: remove sendmail_wait
This is the default anyway.

Note that I can't use the faster -1 (send in the background) setting
with msmtp and Gmail:

http://unix.stackexchange.com/questions/240885/can-i-send-email-in-the-background-in-mutt-while-still-being-notified-of-errors

msmtp/mutt will silently swallow any errors. This is bad enough talking
to an Exchange server (because we at least have a copy of the mail in
the Sent folder, thanks to `set record = ...` and `set copy = yes`), but
for Gmail the message is simply gone.

Thinking about creating a local-only "Outbox" folder and potentially
copying there for Gmail, then using terminal-notifier to let me know
when there is something stuck there. Would need to figure out a way to
clean it up on successful send though. Either way, would probably need
some nasty low-level Message-ID grepping. I wonder if there are any
utils out there for programmatically manipulating maildirs.