mutt: switch from "contacts" to lbdb
"lbdb" provides the same functionality as "contacts" but additionally
can complete based on previous message senders and also aliases.

There is a bug in the Homebrew-installed version, however, which I have
proposed a fix for here:

  https://github.com/Homebrew/homebrew-core/pull/7759

And also reported upstream:

  https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=847735

In any case, my local version was installed manually from source as
described in that PR.

Note that as per:

  https://github.com/tgray/lbdb/issues/5

We need GNU `sed` to be in the `$PATH` in order for the "m_muttalias"
module to work.

You also need something like this to populate the "m_inmail" database:

  find $HOME/.mail/Work/Work -type f -mtime -30d -print -exec sh -c 'cat {} | lbdb-fetchaddr' \;
  find $HOME/.mail/Home/Home -type f -mtime -30d -print -exec sh -c 'cat {} | lbdb-fetchaddr' \;

Still deciding whether to stick that in a cron job of make it an
`offlineimap` postsync hook.