mutt: notify if sync.sh itself dies
We have notifications inside the script already in case one of the
commands dies, but we don't have a notification if the whole thing dies
unexpectedly (due to `set -e`).

This may result in double notifications, but I want to try it out
because I haven't been getting all the notifications that I would expect
when this thing dies.