mutt: use gtimeout (from "coreutils") to catch hung mbsync processes
I've seen these hang for long periods too. Start with a default of 600s
(10 minutes). It remains to be seen how this will interplay with system
sleep; I don't want it to kill jobs that start running shortly before
putting the system to sleep, so may have to figure something else out.

[Edit: Just did a quick test with `gtimeout 60 sleep 30`, put the
machine to sleep for about 90s, woke it up, and it exited without error,
so I think we're good.]