mutt: send HTML email generated from Markdown
Two main approaches to doing this seem to be out there:

(1) Inside mutt, filter the message through `pandoc`, then change the
    MIME-type. For convenience, this can be done with a macro:

    http://unix.stackexchange.com/a/108809/140622

(2) Wrap the external `sendmail` process to transparently add a
    `text/html` part on the fly, again via `pandoc` or similar:

    https://github.com/Roguelazer/muttdown

Each of these sucks in different ways. The big problem with (1) is that
it actually *replaces* the plain-text part, which seems to be throwing
out the baby with the bath water.

The issue with (2) is that the transformation is happening outside of
Mutt, so any copy of the message saved to the sent mail folder by mutt
won't reflect the modification. In the case of Gmail, this is not a
problem, because we don't ask mutt to save those (Gmail does, and it
eventually syncs back), but it is an issue with Office 365, where there
is no option to get it to automatically save.

Overall, though, I think (2) is the better approach, so I cooked up a
Ruby wrapper script to do it. Could have used muttdown, but I don't
understand Python dependency management as much as I do Ruby's, and
also it's fun to tinker and write your own. Probably bugs in there, but
it handles my basic manual testing for now.