vim: lazy-load undotree on first load
Good candidate for lazy loading because the whole plug-in lives in
"plugin/" with no "autoload/" component (at the same time, even though
it's about 1.3k lines of VimL, it still loads fast enough to barely
register in a profiling run).

This is just a first cut, with a couple of problems:

- The "after/plugin/undotree.vim" file executes immediately at startup,
  prior to lazy loading (nevertheless, it still works).
- The ergonomics suck, seeing as you have to escape using `<lt>` when
  specifying a mapping.

I'll improve this subsequently.