vim: use "par" to re-wrap text
Originally started looking into this for mutt, but as noted in the
comments inline, it has some unfavorable interactions there that I can't
be bothered working through at this time. Specifically, it causes a
prompt to appear when auto-viewing HTML email (could probably fix that
with a tweak to the mailcap file, but truth is I only want the
`$display_filter` to run on plain text email parts anyway), and it also
messes up the index display, leaving lingering artifacts.

That means we can't use it as a mutt `$display_filter`, but we can
still use it in Vim, which means that we can reformat ugly mails at
composition time quite easily if we want with `gq` (also `gqq`, although
that is slightly less useful). This is nice because we still have access
to `gw` (and `gww`) to use Vim's internal formatter, in the event that
`par` should ever do the wrong thing.

Using the basic settings recommended in `man par`, but adding `_0`
(digits) to the "body" characters, because I can't think of a situation
in which I'd want these to be considered part of a prefix or suffix.

To avoid a slow `executable` check, we defer setting `'formatprg'`,
except when we're booting directly into composing a mail. This is
probably stupid overkill and I will remove it. I bet `executable()`
introduces an imperceptible delay, and I only did this because I was
thinking of the cost of *forking* `par`, something that we're not doing.

Another drawback to note: `'formatprg'` is not `'textwidth'`-aware. I
think I could hack something together with autocommands, but I will
leave that for another day. `'formatexpr'` is a non-starter because it
runs so often, so I can't fork there, and maintaining cursor position
would be a mess.