Zsh: add `color -` to switch back to previous color
If I am checking out a bunch of color schemes, I often find myself
switching to a theme for a bit and then wanting to switch back to the
one I was previously using. This commit adds this feature to the color
script by teaching it to respond to the special "-" argument. This works
similarly to how other commands like `cd -` and `git checkout -` work,
so it feels pretty natural.

I considered writing the previous color as metadata to my existing
base16 config file to avoid adding another file, but I decided that this
approach would be a little easier to implement so I am rolling with it
for now. But in the future, that might be a possible avenue for
improvement.

I originally did this in
https://github.com/lencioni/dotfiles/commit/249de701 and thought you
might be interested.