brew: don't install Ruby
It is much easier to keep a consistent version of Ruby across Vim,
MacVim, my Mutt set-up etc if I just use the system Ruby for everything.
There is basically no interesting reason for me to use the bleeding edge
for anything at this point.

Not removing `chruby` or `ruby-install` yet, but they probably should go
too. I just blew away `~/.rubies/*` in order to get all of my tools
working correctly with a single version of Ruby. I can pretty much
ignore everything that be relying on having some other version installed
(and fix it to work with the system Ruby if that is the case).