mutt: deal with HTML email without explicit charset
Some HTML mail is well-formed, other HTML mail isn't. Most of the stuff
I've seen is UTF-8 encoded, at least according to mutt, but if you feed
it to Chrome without adequate charset declarations you may see garbage
(example: Todoist digest emails containing emojis).

The "fix" here is an utterly egregious hack. We just *prepend* a `<meta
charset="..."/>` tag to the mail body before handing it off to Chrome
whenever we have a charset (which is basically always): we don't bother
trying to parse potentially invalid HTML or inserting the tag at the
right place inside the `<head>` or even checking for a pre-existing
charset declaration. This is enough to get Chrome to do the right
thing, it doesn't seem to have any nefarious side effects, and it's the
simplest thing that could possibly work (given that we are in a shell
script and we don't want to assume too many dependencies).