mutt: only do work sync on corp network
If we appear to be on the VPN, or corp Wi-Fi, then allow the sync to
run, otherwise don't even bother.

This is very crude, but let's give it a shot. When I am in at work I'll
see what checks I need to implement to detect being wired in to the corp
network.

Encrypting the file because I don't want to leak the name of the
corporate network, but the implementation is basically:

    WIFI=$(networksetup -getairportnetwork en0 | awk '{ print $NF }')

    if [ "$WIFI" = $CORP_SSID ]; then
      echo 'On $CORP_SSID: allowing sync.'
      exit 0
    elif ifconfig utun2 &> /dev/null; then
      echo 'On VPN: allowing sync.'
      exit 0
    else
      echo 'On neither $CORP_SSID nor VPN: disallowing sync.'
      exit 1
    fi