brew: drop flux
I just noticed it isn't working. Did some troubleshooting. Uninstalled,
reinstalled. Deleted prefs. Only then did it start working.

But then I learned that Night Shift is built in on 10.12.4. By all
accounts, it is not as good as Flux and is far more moderate in its
adjustments. Nevertheless, I like the idea of reducing my dependency
foot print, so out it goes.