vim: add "smart" sorter for deoplete
"smart" is a total lie. This is dumb, and barely better than a naive
lexicographic sort. I did this as an exploration of whether you could
get something reasonable without having to implement the full Command-T
algorithm (spoiler alert: you can't, or at least I didn't on this first
try).

The idea:

- Score equals sum of distance between matching letters (no weighting,
  very simple).
- Want earlier matches to be favored, but exact matches later in the
  string should still score well.
- So, ignore distance to first character.
- Use distance to first character only to break ties of equal scoring
  items.

We do this simply in Python by computing a float score where the
integral part is the distance-based-score, ignoring distance to first
character, and the fractional part serves as a tie-breaker by encoding
the distance to the first character.

Great idea, but it only works with a fully recursive dynamic programming
approach, like the one in Command-T, and which I haven't bothered to
implement here. Without this, we do a terrible job of scoring a search
like "glh" on an email like "Greg Hurrell <glh@fb.com>":

  [G]reg Hurre[l]l <gl[h]@fb.com>    0 + 10 + 6 . 0 (ie. 16.0000)

instead of:

  Greg Hurrell <[glh]@fb.com>        0 + 1 + 1 . 14 (ie. 2.0014)

So, back to the ol' drawing board. Trying to figure out the cheapest way
to do this without reimplementing Command-T in Python, or having to do
IPC, or run a daemon, or rebuild Command-T as a Python extension etc.
Maybe there is a decent heuristic.