mutt: add "|" binding for saving (piping) messages
Just wanted to save a raw message, including headers etc, to a file.
According to:

    https://unix.stackexchange.com/a/60891/140622

it should be possible to do this with "|", but I found no such binding
defined by default.

Usage: Hit "|" and at the prompt enter something like:

  cat > ~/Desktop/some-file.txt

Might be nicer to come up with a more convenient binding for this that
instead runs a script and maybe assumes a default message name, but this
will do for now.