zsh: silence annoying ulimit warnings
    ulimit: value exceeds hard limit

Note that the `launchctl` stuff is still working. I can set arbitrary
limit with commands like:

    sudo launchctl limit maxfiles 65535 65535

And see them reflected:

    sudo launchctl limit
    sysctl kern.maxfiles

But it doesn't matter; anything higher than `ulimit -n 10240` prints the
annoying error message.

Best answer for this is still:

    https://unix.stackexchange.com/a/221988/140622

Alas, I don't know why the hard limit kicked in (it started a day or
three before upgrading to High Sierra, so it's not that).

> If the system doesn't let you set the limits above a certain value...
>
> The system doesn't let you set a value higher than a "hard maximum"
> (proposed by Apple). To increase this "hard maximum", you have to
> purchase "OS X Server" from "App Store", then you have to execute the
> following command once:
>
>     sudo serverinfo --setperfmode true
>
> This activates "server performance mode" on your machine. You can
> then set the maximum according to the configuration of your machine
> (see this). I tried this before (on Mountain and Mavericks) and it
> works!