cron: Make `check-git pause` actually work
Never really tested it before, so I hadn't realized it wasn't working.
`lockrun` doesn't use the presence of the lockfile to do the locking; it
actually uses filesystem locking (presumably `flock()`), so the
`rm`/`touch` stuff is ineffective.

A separate problem that I have to troubleshoot is why this cron job,
which I have scheduled to run a few times a day, is running every few
minutes... Ugh.