iTerm: use dynamic profiles to change font size on external monitor
You can't set different font sizes for different screens on macOS, and
that is annoying because the PPI on my external display isn't as high as
my laptop's internal "retina" display:

- External is a DELL UP3216Q: 3840 x 2160 @ 60 Hz (138 PPI)
- Internal: 2880 x 1800 Retina (220 PPI)

As a result, the text looks smaller on the external display.

So, let's steal the idea from here:

https://spatof.org/blog/2016/7/osx-switch-font-size-when-an-external-display-is-attached/

and use Hammerspoon to take advantage of iTerm's "Dynamic Profiles"
feature to update the font size whenever we plug in the external
monitor:

https://www.iterm2.com/documentation-dynamic-profiles.html

This works relatively well because I only ever use the terminal on the
"main" screen, which means that I don't have to worry about the font
looking too big when the screen is on the internal display (it is only
ever there if the external monitor is *not* plugged in).

Chose 16pt as the size because that makes the font look pretty much the
same size as 13pt does on the Retina.

Bonus: this gets a bunch of previous-not-tracked iTerm config under
version control, which is nice, although the truth is I could afford to
omit at least the color-related stuff. Another bonus is we no longer
have any duplication between the default "Wincent" and the "Mutt"
profiles, because we only need to provide overrides for what is
different.

Some follow-up items noted inline.