iterm: symlink instead of copying profile files
Don't want to make the mistake of editing one of the copied files and
then having it blown away later. I want changes to the files to be
reflected in the `git status` output in this repo.

Note that editing the contents of the linked files will almost certainly
not be picked up by iTerm (it is presumably watching the filesystem for
changes to the files in the "DynamicProfiles" directory, but not to
changes in the files that any symlinks may be pointing to). This is ok,
because the change we really care about (switching the symlink to point
at a different profile) *is* picked up.