vim: prevent 'spell' setting from getting blown away in Markdown files
Broken since the addition of an `:ownsyntax` call in a951ca592fcbc15.

Couldn't for the life of me figure out where this was getting
overridden. See the long list of trouble-shooting techniques I went
through here:

    https://vi.stackexchange.com/q/15462/10084

Eventually tracked it down with a manual config bisection (first
eliminated all plugins, then whole files, then lines within files). The
answer is described here:

    https://vi.stackexchange.com/a/15463/10084

As the help notes:

    Note: This resets the 'spell', 'spellcapcheck' and 'spellfile' options.

Now that is a surprise. (Also, seems wrong, as it doesn't mention
'spelllang', which also gets reset.)

One red herring to note here: in my troubleshooting I found that "mail"
files worked but Markdown ones didn't. Turns out that was because `&ft`
was not set for the mail files (I was creating a blank buffer and then
doing `:set ft=mail`), which is why the `:ownsyntax` call wasn't working
and it only seemed to work. Note that I *did* remember this being an
issue inside Mutt, but the fact that it didn't happen in my artificial
test environment didn't really have any rational explanation for me
until it was too late...