karabiner: get usable "fn" and "power" keys on external keyboard
Back in the Karabiner (not Karabiner-Elements) days I used to use "F13"
(the "Print Screen" key) as a sticky "fn" modifier and "F15" (the
"Pause" key) as a substitute for "Power".

Still can't really do sticky modifiers in Karabiner-Elements, but I
realized I have an unused modifier key on the right side of the keyboard
that I can turn into "fn", so doing that here, and setting up the
"pause" -> "power" mapping as well (which seems to map to "sleep").