karabiner: turn Karabiner config into an actual Ansible template
This is pretty ugly (spawning a JS executable, capturing its standard
out to a variable, and then dumping the variable into a template), but
it at least integrates the process for updating this config file with
the process that is used for all the other dot files (ie. to update, you
can test a manual edit to the `~/.config/karabiner/karabiner.json` if
you want, but to actually do it definitively you make a change to the
script and then run `./install dotfiles`).