iterm: hoist variables up to top level
In 50256f3f60b5cc25fb I said that I didn't like the implicit dependency
of the "dotfiles" role on the "iterm" role (for the purposes of getting
these variables). Just got bitten by it (I had run the "dotfiles" role
and forgotten to run the "iterm" role beforehand), so moving the
variables up into the top level, lest this happen again.