chrome: Increase zoom on 4k display
Idea via this tweet:

https://twitter.com/LukasReineke/status/993972191329964033

Use a media query to target low-DPI displays and boost `zoom`. I picked
a value of 1.3 because it seems to be a pretty good match in terms of
the resulting physical font size. Some resized images do look a bit
blurry, but then again, everything looks a bit blurry on a non-Retina
display.

Nice thing about this method is that browser windows will update their
zoom immediately as soon as you move windows from one display to
another.

In a previous commit I stated that my displays have PPIs of 138 and 220
respectively:

- External is a DELL UP3216Q: 3840 x 2160 @ 60 Hz (138 PPI)
- Internal: 2880 x 1800 Retina (220 PPI)

This is mostly confirmed by http://dpi.lv/ which claims they are 140 and
220 respectively.

But Chrome evidently thinks my Retina has a DPI of 192, because that's
where the media query kicks in. Oh well, whatevs. ¯\_(ツ)_/¯

I'm using max-resolution here because I don't want to set a default zoom
of 1 and then override it with a `min-resolution` or exact `resolution`
media query. The one-off-ness here is a bit ugly, but it works.