Revert "vim: Maintain VISUAL selection while indenting/dedenting"
This reverts commit 96580a53960afd72e295d728062ed21f445c9999.

In that commit I wrote:

    I can't remember if I've tried this before.

and now I know that I must have and the thing that drove me up the wall
then was that "." no longer works. IIRC I tried to get repeat.vim to
make this work, but it doesn't. In fact, a little digging links me back
to this 2015 issue, filed by me:

    https://github.com/tpope/vim-repeat/issues/48