vim: fix spell-checking insanity
Something changed either with a Vim update or a plug-in update which has
broken spell-checking for me in both Neovim (v0.3.1) and Vim (v8.1).

I complained about this a couple of months back:

https://www.reddit.com/r/vim/comments/8yrsxv/spell_checking_suddenly_went_haywire_help/

Basically, I'd find that any time I added a word to my word list with
"zg", Vim/Neovim would create:

  ~/.vim/spell/en
  ~/.vim/spell/en.utf-8.spl

instead of adding to my word list at:

  ~/.vim/spell/en.utf-8.add

On next launch, Vim/Neovim would see the ".spl" file as the dictionary
because it is presumably at a place that comes first in the runtime
path, meaning that everything would be underlined as misspelt.

For reference, this is where the official files are:

  /usr/share/vim/vim80/spell/en.utf-8.spl (Apple's Vim)
  /usr/local/Cellar/vim/8.1.0001/share/vim/vim81/spell/en.utf-8.spl
  /usr/local/Cellar/neovim/0.3.1/share/nvim/runtime/spell/en.utf-8.spl

You could "fix" this by deleting the unwanted files, but they would just
come back again on the next "zg".

The other issue is that only Vim, not Neovim, was consulting the
preexisting ".add" file.

This fix here is to set "spellfile". As far as I can tell, hard-coding
this to be "en.utf-8" is fine because Vim is going to use the first lang
in "spelllang" anyway (which is "en" in my case).

Tested this in Neovim and Vim: can add words to the file, and it
continues to approve of both English and Spanish words.