Force some things to run in --check mode
Because of the Ansible upgrade I am running all the roles with `--check`
to see what issues arise. Unfortunately, if we `register` a `command` in
`--check` mode, we wind up not actually running, so the result object
has no `rc` key, which then leads to an error like this when you try to
access it:

    The conditional check 'item.1.rc == 0' failed. The error was: error
    while evaluating conditional (item.1.rc == 0): 'dict object' has no
    attribute 'rc'

    The error appears to have been in
    '/Users/glh/code/wincent/roles/dotfiles/tasks/main.yml': line 25,
    column 3, but may be elsewhere in the file depending on the exact
    syntax problem.

    The offending line appears to be:

    - name: dotfiles | backup originals
      ^ here

The commands here are harmless, so let's just run them in `--check` mode
by sitting `check_mode: no`. I could probably also looking at using the
"stat" module instead for at least the first of these use cases.