vim: follow LanguageClient-neovim "next" branch
In the last commit, I mentioned I wasn't sure which branch to follow
oup of "next" (one commit ahead of latest release), "dev" (currently
pointing at latest -- ie. v0.1.120 -- release) and "master" (very
stale).

Spelunking with `gitk`/`git` (or `git gr --all`, an alias) showed a
not-at-all illuminating commit many months back when the "next" branch
was forked, but a latter commit that updated the README and showed that
the "next" branch is the "blessed" one.

So, I'm switching to that.