vim: Hack in basic Goyo support
I checked out a few "distraction-free writing" plug-ins for Vim because
my Zsh `wr` function isn't really cutting it:

- https://github.com/amix/vim-zenroom
- https://github.com/vim-scripts/WriteRoom-for-Vim
- https://github.com/mikewest/vimroom
- https://github.com/junegunn/goyo.vim

Of these, Goyo seems the best. I have a bunch of autocmd stuff going on
that gets in the way, so I am hackily turning it off and on again for
the purposes of this trial. If I like it, I will try to polish it up a
bit.

For best results, do `:Goyo` in a Vim instance running in a dedicated
terminal window (no tabs).