karabiner: replace hard-coded mapping with JSONPath-based modification
This is almost pointless, but I didn't like the hard-coded-ness of the
previous commit. Here we make a tiny JSONPath traverser (supporting only
a small subset of the possible selection types) that we use to update
our "conditions" field).

Because this moves complexity out of our previous custom traversal (into
the abstraction provider by the `visit()` function) we are now free to
make out modifications ever so slightly more sophisticated, without
compromising readability: specifically, we skip adding the
"frontmost_application_unless" condition if we detect a
"frontmost_application_if" block is already there.