zsh/tmux: tweak placement of NBSP in prompt
Seeing as tmux's word-separators option doesn't work with UTF-8, we
can't yet at NBSP to it. But if we move the NBSP earlier in the prompt
(unfortunately, making it one characters wider, seeing as ZWSP -- the
zero-width space -- isn't viable at this time) then we can have a real
space before the start of the command, which means we can add a
`next-word` to our "last command" binding, which is a slight usablility
improvement.