vim: add CommandTBoot helper command
This allows me to launch a tmux and immediately fire off a Vim session
with a Command-T prompt ready to go without any glitches. You get
glitches if you do this:

  tmux send-keys -t $TARGET "vim -c CommandT" Enter

Depending on the session, you can kludgily work around the glitches
with:

  tmux send-keys -t $TARGET "sleep .75; vim -c CommandT" Enter

but that is annoying and obviously pretty fragile.

By defering the command until post-VimEnter, the glitches go away:

  tmux send-keys -t $TARGET "vim -c autocmd VimEnter * CommandT" Enter

but that is verbose and also leaves a dangling autocmd around. With this
commit, we can instead do:

  tmux send-keys -t $TARGET "vim -c CommandTBoot" Enter