zsh: tweak history options
Turn off HIST_IGNORE_ALL_DUPES because it prevents me from accurately
examining my most frequently used commands.

I don't know if it defaults to on or not (it is not mentioned in the
defaults section of the manual) although I do know that the related
HIST_IGNORE_DUPS defaults to on, so let's just turn them both off
explicitly.

Turn on HIST_FIND_NO_DUPS, which also is not documented as defaulting to
on or off, but in practice sure seems to be on; turn it on anyway to be
sure. It gives me the behavior that I want: full history recorded to
`$HISTFILE`, but duplicates avoided when navigating history via
readline.