tmux: add colon to default list of word separators
Because often paths are printed with a trailing line number (eg.
filename.ext:100) and what we usually want to do is select just the
name; if we want the number too, it is easy to extend the selection and
grab it (easier than reducing the selection the way things were
previously configured).