Fix Ansible template bustage
This was inadvertently introduced in 62f5656fd5779 because I edited the
templated file in place and then copied it back to the source; naughty!

The change was causing Ansible to complain thusly:

```
AnsibleError: template error while templating string: Encountered
unknown tag 'b'
```

Rather than rolling back the fix as it was before, I am going for what
is a hopefully more readable version by putting more context in the
interpolated string, as opposed to just the problematic sequence ({%b).