keyboard: turn down key repeat before going to lock screen
As mentioned here:

    https://twitter.com/wincent/status/1068234903811223552

Going to see if this can cure my lock screen woes. You can turn off key
repeat with:

    defaults write -g KeyRepeat 300000

But you have to log out for it to take effect. I did some reverse
engineering and header-file searching to figure out how:

    /System/Library/PreferencePanes/Keyboard.prefPane/Contents/MacOS/Keyboard

Is forcing the changes to take effect immediately, and discovered some
deprecated (sigh) APIs that hopefully will continue to exist for a
while:

    NXSetKeyRepeatInterval()
    NXKeyRepeatInterval()
    NXOpenEventStatus()

I bundled these up in a `dry` executable (short for "Don't Repeat
Yourself") that can be used to set the key repeat delay in seconds. Not
building or installing this automatically yet so that I can test it.

Calling `dry` from the `karabiner-kill` and `karabiner-boot` scripts,
and also switched to from running on systemWillSleep to screensDidLock,
which allows us to cover both normal screen saver and sleep events with
a single branch.