vim: stop vim-javascript plug-in from clobbering jest filetype
Not sure when this started: either the plug-in has changed or we have a
load-ordering issue.

In any case, be ready for it and swoop in whenever the filetype is set
to "javascript". If "jest" is already in there, nothing to do. If it
isn't, make sure the filename pattern is right and if so, stick "jest"
(back) on the filetype.