zsh/iTerm: rework tab/window titles
Somehow this had stopped working, even though Zsh was setting the title
and tmux was setting the title too (incorporating the title set by Zsh).
In the iTerm settings I turned messed with the "Title" setting in the
Session tab (Command-I), setting it to "Session Name" (shows up as
"Name" in the closed pop-up), and that seemed to let the underlying
title come through again. I've pulled across what I think may be the
corresponding setting into 00-Base.json.

Then at the tmux level I changed it to set the title to:

    [tmux session name] > [whatever zsh wants the title to be]

Previously it was:

    [whatever zsh wants the title to be] > [short hostname] > [tmux session name] > [tmux window name]

In practice, I usually know exactly what host I am on, the session name
is most important for identifying which tab to switch to, and the window
name is next to useless.

At the zsh level I changed it from:

(1) At the prompt:
    [last command, with all args]
(2) While running a command:
    [current command name is, with no args]

to the following when outside tmux:

(1) At the prompt, in new shells:
    [current working directory] (because the last command is meaningless)
(1b) At the prompt, in non-new shells:
    [current working directory] > [last command name, no args]
(2) While running a command:
    [current working directory] > [current command name, no args]

and this when inside tmux:

(1) At the prompt, in new shells:
    [current working directory] (because last command is meaningless)
(1b) At the prompt, in non-new shells:
    [last command name, no args]
(2) While running a command:
    [current command name, no args]

In other words:

- We always show some context (current working directory outside tmux
  and tmux session name inside it).
- We consistently only show command names and not sometimes commands +
  args (in generally the extra info is excessive and distracting).

Putting it all together, you always see:

  [context] > [last/current command name]

or, if there is no last or current command to show:

  [context]