karabiner: disable self with Right-Shift x 3
So my keyboard just jammed at the lockscreen again:

    https://wincent.com/snippets/1162

Seems the Hammerspoon sleep hook did not finish running because
Karabiner was active (I could toggle caps lock with both Shift keys
tapped together) and key repeat was still on.

So, teach Karabiner to turn itself off with Right-Shift x 3. (If this
actually works, I will DRY up the config later on.)

I found that `karabiner-kill` doesn't work when called from Karabiner
itself. It dies silently after the first launchctl call, which seems to
work (because the user console server is indeed killed). No errors
logged anywhere. I thought it might be getting killed when its parent
exited, but even running with a `nohup ... &` wrapper doesn't fix that
(and I can see in pstree output that the script is owned by launchd
itself), so something else is going on. I suspect it might be because
it may be running in a non-GUI context so something fails silently and
hard.

Hack to make this work is to instead get `osascript` to run an
AppleScript telling the Terminal to actually do the script running:

```
-- Workaround to enable karabiner-kill to be called from a Karabiner shell_command:
-- https://github.com/tekezo/Karabiner-Elements/issues/1573
tell application "Terminal"
	activate
	do script "/bin/sh ~/.zsh/bin/karabiner-kill"
end tell
```

Hideous, I know, but if it saves us from doing hard resets it will be
worth all the ghastliness..