karabiner: switch kill binding
I realized that Right-Shift x 3 isn't going to work when the keyboard is
jammed repeating, because any repeat is going to cause the
to_delayed_action to immediately go to to_if_canceled.

So, instead go for Fn+Control+Option+Command+Z. Z is the same on both
Colemak and Qwerty layouts. Tested this by holding down X and then
pressing the binding. It works.