vim: fix some more ColorColumn clashes (in Git commits)
If you look at `:verbose hi DiffNewFile` you see a bunch of these, many
with backgrounds set, which will manifest the same problem if the lines
are long. But I don't know whether we'll see those long lines in
practice, so only doing it for the ones that set the background color to
black (the three groups that definitely tend to have long lines are
`DiffAdded`, `DiffFile` and `DiffRemoved`; the other two here, I'm not
sure about).