vim: use Flow in LSP mode for LanguageClient-neovim
I am pretty sure I am going to tweak this further but this is an initial
start. Main change I will want to make is to not set the bindings
unconditionally, but in specific buffers, I think (for example, it is
still useful to have `K` work in Vimscript files, or `gd` work in a
JavaScript file which doesn't have a Flow config.

In order to make that happen will likely need to employ a heuristic: if
the $PWD has a `.flowconfig`, then we'll set up the server commands. Not
perfect, but it will get what I want 99% of the time (I rarely `:cd`).
Then we'll need some kind of buffer-local bindings based on the
registered server commands.