vim: use location list for diagnostic warnings
The plug-in defaults to using the quickfix listing, which is a bit
aggressive (blows away contents as you type, naturally causing a bunch
of intermediate invalid states). Maybe even using the location list is
too aggressive, as I do use that for other things from time to time. I
may end up making this one run on-demand instead of automatically.